moved font.c to font.h
